+++ modelicafmt_test.go
@@ -32,13 +32,15 @@ var exampleFileTests = []struct {
 	outFile         string
 	formatterConfig Config
 }{
-	{"gmt-coolingtower.mo", "gmt-coolingtower-out.mo", Config{-1, false}},
-	{"functions.mo", "functions-out.mo", Config{-1, false}},
-	{"example-no-within.mo", "example-no-within-out.mo", Config{-1, false}},
-	{"example-arrays.mo", "example-arrays-out.mo", Config{-1, false}},
-	{"gmt-building.mo", "gmt-building-out.mo", Config{-1, false}},
-	{"gmt-building.mo", "gmt-building-80-out.mo", Config{80, false}},
-	{"gmt-building.mo", "gmt-building-empty-lines-out.mo", Config{-1, true}},
+	{"gmt-coolingtower.mo", "gmt-coolingtower-out.mo", Config{-1, false, false}},
+	{"functions.mo", "functions-out.mo", Config{-1, false, false}},
+	{"example-no-within.mo", "example-no-within-out.mo", Config{-1, false, false}},
+	{"example-arrays.mo", "example-arrays-out.mo", Config{-1, false, false}},
+	{"example-arrays.mo", "example-arrays-wrapped-out.mo", Config{-1, false, true}},
+	{"example-arrays-nd.mo", "example-arrays-nd-out.mo", Config{-1, false, true}},
+	{"gmt-building.mo", "gmt-building-out.mo", Config{-1, false, false}},
+	{"gmt-building.mo", "gmt-building-80-out.mo", Config{80, false, false}},
+	{"gmt-building.mo", "gmt-building-empty-lines-out.mo", Config{-1, true, false}},
 }
 
 func TestFormattingExamples(t *testing.T) {
@@ -82,7 +84,7 @@ func TestProcessFileReturnsErrorOnInvalidInput(t *testing.T) {
 	defer os.Remove(sourceFile)
 
 	var out bytes.Buffer
-	err := processFile(sourceFile, &out, Config{-1, false})
+	err := processFile(sourceFile, &out, Config{-1, false, false})
 
 	a.Error(err, "processFile should return an error for input with unrecognized tokens")
 }